+++ veltahleon/src/main/java/com/testament/veltahleon/services/ifc/society/PersonDetailsService.java
@@ -6,5 +6,6 @@ public interface PersonDetailsService {
 
     PersonDetails getPersonByFirstNameAndSecondName(String firstName, String secondName);
     Boolean deletePersonDetailsByID(Long id);
+    PersonDetails updatePersonDetails(Long id, PersonDetails personDetails);
     PersonDetails modifyPersonDetails(Long id, PersonDetails personDetails);
 }